QroTravel/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.6
django-ckeditor==5.7.1
django-cors-headers==3.1.0
django-debug-toolbar==2.0
django-geoposition==0.3.0
django-js-asset==1.2.2
django-modeltranslation==0.13.3
django-robots==3.1.0
django-solo==1.1.3
django-stdimage==5.0.0
django-translatable-fields==0.0.1
facebook-sdk==3.1.0
get==2019.4.13
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mod-wsgi==4.6.2
mysqlclient==1.4.4
Pillow==6.1.0
post==2019.4.13
progressbar2==3.42.0
public==2019.4.13
pylint==2.3.1
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
twitter==1.18.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
